Add autodetected http/https protocols
This is in response to @tpope's response on PR #590 to allow https to be
used for localhost connections using Preview.  There, he mentioned that
he'd like the uri scheme protocol to be autodetected, which I
appreciate.  I opted to use 'possible approach 3' to ps/grep for
ssl://<binding> to see if it was being used or not.